New moment system

+++ src/components/blocks/spike.tsx
@@ -1,14 +1,31 @@
-import { useContext, useState } from "react";
-import { Player, PlayerContext } from "@/context/playercontext";
-export default function Spike() {
-  const player = useContext(PlayerContext);
+import { useContext, useState, useEffect } from "react";
+import {PlayerContext } from "@/context/playercontext";
 
-  const [isDead, setIsDead] = useState(false);
-  if (player) {
-    if (player.postion.x === 0 && player.postion.y === 0) {
-      player.kill();
-      setIsDead(true);
+export default function Spike(props: {
+  position: {
+    x: number;
+    y: number;
+  },
+  spikeFrequenzy?: number;
+  startState?: boolean;
+}) {
+  const players = useContext(PlayerContext);
+
+  const [isDeadly, setIsDead] = useState(props.startState ?? false);
+
+  useEffect(() => {
+    const interval = props.spikeFrequenzy ?? 5000;
+    setInterval(() => {
+      setIsDead((prev) => !prev)
+    }, interval)
+  })
+
+  if (players) {
+    for (const player of Object.values(players)) {
+      if (player.postion.x === props.position.x && player.postion.y === props.position.y) {
+        player.kill();
+      }
     }
   }
-  return <div style={{ width: "100%", height: "100%", background: isDead ? "red" : "black" }}></div>;
+  return <div style={{ width: "100%", height: "100%", background: isDeadly ? "red" : "black" }}></div>;
 }
